requirements.txt
amqp==2.1.3
anyjson==0.3.3
appdirs==1.4.2
billiard==3.5.0.2
celery==4.0.2
defusedxml==0.5.0
Django==1.10.5
django-celery-beat==1.0.1
django-celery-results==1.0.1
django-crispy-forms==1.6.1
django-debug-toolbar==1.6
django-redis==4.7.0
django-summernote==0.8.6
djangorestframework==3.5.4
geetest==3.2.1
kombu==4.0.2
oauthlib==2.0.1
packaging==16.8
PyMySQL==0.7.10
pyparsing==2.1.10
python3-openid==3.1.0
pytz==2016.10
redis==2.10.5
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0
sqlparse==0.2.2
vine==1.1.3
